+++ packages/caver-transaction/src/transactionTypes/abstractFeeDelegatedTransaction.js
@@ -23,7 +23,7 @@ const Bytes = require('eth-lib/lib/bytes')
 const TransactionHasher = require('../transactionHasher/transactionHasher')
 const AbstractTransaction = require('./abstractTransaction')
 const { refineSignatures, typeDetectionFromRLPEncoding } = require('../transactionHelper/transactionHelper')
-const Keyring = require('../../../caver-wallet/src/keyring/keyring')
+const Keyring = require('../../../caver-wallet/src/keyring/keyringFactory')
 const AbstractKeyring = require('../../../caver-wallet/src/keyring/abstractKeyring')
 const { KEY_ROLE } = require('../../../caver-wallet/src/keyring/keyringHelper')
 const utils = require('../../../caver-utils/src')
