+++ packages/caver-transaction/src/transactionTypes/abstractTransaction.js
@@ -22,7 +22,7 @@ const RLP = require('eth-lib/lib/rlp')
 const Hash = require('eth-lib/lib/hash')
 const TransactionHasher = require('../transactionHasher/transactionHasher')
 const utils = require('../../../caver-utils')
-const Keyring = require('../../../caver-wallet/src/keyring/keyring')
+const Keyring = require('../../../caver-wallet/src/keyring/keyringFactory')
 const AbstractKeyring = require('../../../caver-wallet/src/keyring/abstractKeyring')
 const { TX_TYPE_STRING, refineSignatures, typeDetectionFromRLPEncoding } = require('../transactionHelper/transactionHelper')
 const { KEY_ROLE } = require('../../../caver-wallet/src/keyring/keyringHelper')
