+++ test/packages/caver.transaction/feeDelegatedSmartContractExecution.js
@@ -32,7 +32,7 @@ const { propertiesForUnnecessary } = require('../utils')
 
 const testRPCURL = require('../../testrpc')
 const Caver = require('../../../index.js')
-const Keyring = require('../../../packages/caver-wallet/src/keyring/keyring')
+const Keyring = require('../../../packages/caver-wallet/src/keyring/keyringFactory')
 const SingleKeyring = require('../../../packages/caver-wallet/src/keyring/singleKeyring')
 const TransactionHasher = require('../../../packages/caver-transaction/src/transactionHasher/transactionHasher')
